fix & refactor: gnb 디자인 에러 수정 및 리팩토링
