+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import localFont from "next/font/local"
 import { cookies } from "next/headers"
 
 import QueryProviders from "@/components/app/provider"
-import GNB from "@/components/public/gnb/GNB"
+import RouteValidationLayout from "@/components/public/gnb/RouteValidationLayout"
 import { CountProvider } from "@/provider/CountProvider"
 import ToastProvider from "@/provider/ToastProvider"
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools"
@@ -70,11 +70,11 @@ const RootLayout = ({
   const userToken = cookieStore.get("userToken")?.value
   return (
     <html lang="ko" className={`${pretendard.className} ${tmoneyRoundWind.variable}`}>
-      <body className="bg-gray-100 pt-[56px] md:pt-[60px]">
+      <body className="bg-gray-100">
         <CountProvider>
           <ToastProvider>
             <QueryProviders>
-              <GNB userToken={userToken}>{children}</GNB>
+              <RouteValidationLayout userToken={userToken}>{children}</RouteValidationLayout>
               {process.env.NODE_ENV !== "production" && <ReactQueryDevtools position="bottom" />}
             </QueryProviders>
           </ToastProvider>
